+++ packages/StunDartTests/test/stun_registry_test.dart
@@ -3,7 +3,6 @@ import 'dart:io';
 import 'package:singleton_manager/singleton_manager.dart';
 import 'package:stun/src/initial_point/initial_point_registry.dart';
 import 'package:stun/src/interfaces/i_stun_handler_base.dart';
-import 'package:stun/stun.dart';
 import 'package:test/test.dart';
 
 const _key1 = 'primary';